BIG UPDATED, Reformatted how many of the pages work to use solely the data
