+++ src/routes/Blog.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import Data from './Data/Data.json'
-import Layout from './Data/Designs/Components/Layout.js'
+import Layout from './Data/Applications/Components/Layout.js'
 import './Stylesheets/Blog.css'
 
 
@@ -15,15 +15,15 @@ class Blog extends Component {
 					Posts
 					<hr style={{width:"100%"}}/>
 					<div className="ScrollBox">
-						{Data ? Data.Applications.map((item)=>{
+						{Data && Data.Blogs ? Data.Blogs.map((item)=>{
 							if(item.Image === null) item.Image = "/Images/WIP.png"
 							return(
-								<a key={Data.Applications.indexOf(item)} className="Link" style={{backgroundImage:"URL("+item.Image+")"}} href={document.location.origin+"/Applications/"+item.Name}>
+								<a key={Data.Blogs.indexOf(item)} className="Link" style={{backgroundImage:"URL("+item.Image+")"}} href={document.location.origin+"/Blog/"+item.Name}>
 									<div className="Info">
 										<div className="Name">{item.Name}</div>
 										<hr/>
 										<div className="Date">{item.Date}</div>
-										<div className="Description">{item.Description}</div>
+										<div className="Description">{item.Paragraphs[0]}</div>
 									</div>
 								</a>)
 						}) : <div>No Data</div>}
