+++ src/routes/Data/Applications/Components/NavBar.js
@@ -5,6 +5,7 @@ class NavBar extends Component {
 		return (
 			<span className="NavBar">
 				<a className="Link BoldLink" href={document.location.origin+"/Contact"}>Contact</a>
+				<a className="Link" href={document.location.origin+"/Blog"}>Blog</a>
 				<a className="Link" href={document.location.origin+"/Designs"}>Designs</a>
 				<a className="Link" href={document.location.origin+"/Applications"}>Applications</a>
 				<a className="Link" href={document.location.origin}>Home</a>
