+++ src/routes/NotFound.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import Layout from './Data/Designs/Components/Layout.js'
+import Layout from './Data/Applications/Components/Layout.js'
 import './Stylesheets/NotFound.css'
